requirements.txt
fastapi
uvicorn
pandas
numpy
matplotlib
seaborn
requests
beautifulsoup4
openai
PyPDF2
Pillow
duckdb
python-multipart
aiofiles
httpx
# requirements.txt
fastapi
uvicorn
httpx
pandas
numpy
matplotlib
beautifulsoup4
lxml
duckdb
Pillow
aiofiles
torch
transformers
accelerate
# Universal Data Analyst Agent - Requirements
# Core web framework
flask
flask-cors

# Google Gemini AI
google-generativeai

# Data processing and analysis
pandas
numpy
scipy
openpyxl

# Visualization
matplotlib
seaborn

# Web scraping
requests
beautifulsoup4
selenium

# File processing
PyPDF2
Pillow
python-docx

# Database support
duckdb

# Configuration
pyyaml
toml

# Additional utilities
python-dotenv
lxml

# Optional: For better async support
aiohttp
aiofiles
